==> final/07_sample_hyperparameter_search_template.ipynb <==
import optuna
from functools import partial
from src.objective import objective


# Create the model for example
# model = UserEmbeddingModel(input_dim=input_dim,
#                            embedding_dim=embedding_dim,
#                            num_users=num_users,
#                            dropout_prob=0,
#                            learning_rate=lr,
#                            lr_gamma=lr_gamma,
#                            l2_reg=l2_reg)
model = ...

#Load the dataset
train_dataset = ... 

objective_with_data = partial(objective, train_dataset=train_dataset, model=model)
study = optuna.create_study(direction="minimize")
study.optimize(objective_with_data, n_trials=20)
print(f"Best trial: {study.best_trial.value}")
print(f"Best hyperparameters: {study.best_trial.params}")
